add curl and openssl

+++ native/hook/exports.cpp
@@ -16,7 +16,8 @@ void EXPORTS_API perfcat_hook_init(perfcat::perfcat_hook_init_t args) {
 
 #ifdef _WIN32
   auto wake_up_tid = args.wake_up_tid;
-  auto wake_up_handle = OpenThread(THREAD_ALL_ACCESS, FALSE, wake_up_tid);
+  auto wake_up_handle =
+      OpenThread(THREAD_ALL_ACCESS, FALSE, static_cast<DWORD>(wake_up_tid));
   if (wake_up_handle == nullptr) {
     return;
   }
